optimize the recyclerview
