+++ RecyclerViewWithHeader/app/src/main/Kotlin/com/tutorial/mim/recyclerview_header/model/ItemHolder.kt
@@ -1,14 +1,13 @@
 package com.tutorial.mim.recyclerview_header.model
 
-import android.content.Context
-import android.util.Log
 import android.view.View
 import kotlinx.android.synthetic.main.list_content.view.*
+import kotlinx.android.synthetic.main.list_header.view.*
 
 /**
  * Created by lf_wannabe on 27/08/2017.
  */
-class ItemHolder(itemView: View) : BaseViewHolder<Item>(itemView) {
+class ItemHolder(itemView: View) : BaseViewHolder<Item, Item>(itemView) {
     override fun onBind(item: Item?) {
         with(itemView){
             number.text = item?.num.toString()
@@ -16,7 +15,10 @@ class ItemHolder(itemView: View) : BaseViewHolder<Item>(itemView) {
         }
     }
 
-    override fun onBind(data: Item, onClickListener: View.OnClickListener, context: Context) {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    override fun onBindHeader(header: Item?) {
+        itemView.header.text = header?.title
     }
+
+
+
 }
